+++ packages/class-core-test/src/value-managers/containment_1.tests.ts
@@ -23,14 +23,13 @@ import {
     ChildMovedFromOtherContainmentInSameParentDelta,
     ChildReplacedDelta,
     collectingDeltaReceiver,
-    nodeBaseDeserializer,
     serializeNodeBases
 } from "@lionweb/class-core"
-import { AccumulatingSimplisticHandler } from "@lionweb/core"
 import { LionWebJsonMetaPointer } from "@lionweb/json"
 
 import { LinkTestConcept, TestLanguageBase } from "@lionweb/class-core-test-language"
 import { deepEqual, equal, isTrue, isUndefined, throws } from "../assertions.js"
+import { deserializeNodesAssertingNoProblems } from "./tests-helpers.js"
 
 const testLanguageBase = TestLanguageBase.INSTANCE
 
@@ -265,10 +264,7 @@ describe("serialization and deserialization w.r.t. a [1] containment", () => {
         const serContainment = nodes[0].containments.find(({containment}) => containment.key === metaPointer.key);
         isUndefined(serContainment);
 
-        const deserialize = nodeBaseDeserializer([testLanguageBase]);
-        const problemHandler = new AccumulatingSimplisticHandler();
-        const deserializedNodes = deserialize(serializationChunk, undefined, undefined, undefined, problemHandler);
-        equal(problemHandler.allProblems.length, 0);
+        const deserializedNodes = deserializeNodesAssertingNoProblems(serializationChunk);
         equal(deserializedNodes.length, 1);
         const root = deserializedNodes[0];
         isTrue(root instanceof LinkTestConcept);
@@ -297,10 +293,7 @@ describe("serialization and deserialization w.r.t. a [1] containment", () => {
             children: ["child"]
         });
 
-        const deserialize = nodeBaseDeserializer([testLanguageBase]);
-        const problemHandler = new AccumulatingSimplisticHandler();
-        const deserializedNodes = deserialize(serializationChunk, undefined, undefined, undefined, problemHandler);
-        equal(problemHandler.allProblems.length, 0);
+        const deserializedNodes = deserializeNodesAssertingNoProblems(serializationChunk);
         equal(deserializedNodes.length, 1); // (because there's only one “root”)
         const root = deserializedNodes[0];
         isTrue(root instanceof LinkTestConcept);